+++ apps/pick-learn/src/shared/assets/icons/Star.tsx
